+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - bievents - add optional state to events
-- bievents - WithTenantID option for emitter
+- bievents - `WithTenantID` option for emitter
+- d4lcontext - `GetUserIDFromCtx`, `GetClientIDFromCtx`, `GetTenantIDFromCtx` to extract information from a context object directly
 
 ### Changed