+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- DB2 pkg - use migrate pkg after the execution of the gorm migrations
-
 ### Changed
 
 ### Deprecated
@@ -20,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v1.13.0] - 2021-08-02
+
+### Added
+
+- DB2 pkg - use migrate pkg after the execution of the gorm migrations
 
 ## [v1.12.0] - 2021-08-02
 
@@ -325,7 +329,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.12.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.13.0...HEAD
+[v1.13.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.12.0...v1.13.0
 [v1.12.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.11.0...v1.12.0
 [v1.11.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.10.0...v1.11.0
 [v1.10.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.9.0...v1.10.0